grupo10/requirements.txt
asgiref==3.8.1
Django==3.2.25
filetype==1.2.0
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
